Give toybox a proper structure
